Move shared application_controller and back_office_forms_controller logic to a controller concern
